Split emit-init-sections.ts into 4 focused modules
Reduces the largest file (1,520 lines, 30% of ir-to-client-js) into
4 files organized by responsibility:

- emit-control-flow.ts (554 lines) — conditionals + loops
- emit-reactive.ts (253 lines) — attribute/text/prop reactive updates
- emit-registration.ts (391 lines) — constants resolution + hydrate
- emit-init-sections.ts (354 lines) — declarations, props, events, refs

No logic changes. Compiled output is byte-identical before and after.
Import paths updated in generate-init.ts and index.ts.
