+++ packages/jsx/src/ir-to-client-js/emit-init-sections.ts
@@ -1,48 +1,16 @@
 /**
  * Extracted output phases from generateInitFunction.
  * Each function appends to a lines[] array.
+ *
+ * Declarations, props extraction, event handlers, refs, and lifecycle.
+ * Control flow, reactive updates, and registration are in separate modules.
  */
 
-import type { AttrMeta, ComponentIR, SignalInfo, IRFragment } from '../types'
+import type { SignalInfo } from '../types'
 import type { Declaration } from './declaration-sort'
-import { isBooleanAttr } from '../html-constants'
-import type { ClientJsContext, ConditionalBranchEvent, ConditionalBranchRef, ConditionalBranchChildComponent, ConditionalBranchTextEffect, LoopChildEvent, LoopElement } from './types'
-import { inferDefaultValue, toHtmlAttrName, toDomEventName, wrapHandlerInBlock, buildChainedArrayExpr, quotePropName, varSlotId, bodyReferencesComponentScope, PROPS_PARAM } from './utils'
-import { addCondAttrToTemplate, canGenerateStaticTemplate, irToComponentTemplate, generateCsrTemplate, irChildrenToJsExpr, createStringProtector } from './html-template'
+import type { ClientJsContext } from './types'
+import { inferDefaultValue, toDomEventName, wrapHandlerInBlock, varSlotId, quotePropName, PROPS_PARAM } from './utils'
 
-/**
- * Generate JS statements to update a DOM attribute reactively.
- * Centralizes the attribute-type dispatch (value, class, boolean, presence, generic)
- * so that new AttrMeta flags are handled in one place.
- */
-function emitAttrUpdate(target: string, attrName: string, expression: string, meta: AttrMeta): string[] {
-  const htmlName = toHtmlAttrName(attrName)
-  if (htmlName === 'class') {
-    return [
-      `{ const __v = ${expression}; if (__v != null) ${target}.setAttribute('class', String(__v)); else ${target}.removeAttribute('class') }`,
-    ]
-  }
-  if (htmlName === 'value') {
-    return [
-      `const __val = String(${expression})`,
-      `if (${target}.value !== __val) ${target}.value = __val`,
-    ]
-  }
-  if (isBooleanAttr(htmlName)) {
-    return [`${target}.${htmlName} = !!(${expression})`]
-  }
-  if (meta.presenceOrUndefined) {
-    // aria-* requires explicit "true" value per WAI-ARIA spec
-    const attrVal = htmlName.startsWith('aria-') ? 'true' : ''
-    return [
-      `if (${expression}) ${target}.setAttribute('${htmlName}', '${attrVal}')`,
-      `else ${target}.removeAttribute('${htmlName}')`,
-    ]
-  }
-  return [
-    `{ const __v = ${expression}; if (__v != null) ${target}.setAttribute('${htmlName}', String(__v)); else ${target}.removeAttribute('${htmlName}') }`,
-  ]
-}
 
 /**
  * Collect slot IDs that are inside conditionals (handled by insert()).
@@ -228,668 +196,6 @@ export function emitPropsEventHandlers(
   }
 }
 
-/** Emit createEffect blocks that update text nodes for reactive expressions. */
-export function emitDynamicTextUpdates(lines: string[], ctx: ClientJsContext): void {
-  // Group elements by expression to consolidate effects with same dependencies
-  const byExpression = new Map<string, typeof ctx.dynamicElements>()
-  for (const elem of ctx.dynamicElements) {
-    const key = elem.expression
-    if (!byExpression.has(key)) {
-      byExpression.set(key, [])
-    }
-    byExpression.get(key)!.push(elem)
-  }
-
-  for (const [expr, elems] of byExpression) {
-    // Separate conditional vs non-conditional elements
-    const conditionalElems = elems.filter(e => e.insideConditional)
-    const normalElems = elems.filter(e => !e.insideConditional)
-
-    if (normalElems.length > 0 || conditionalElems.length > 0) {
-      lines.push(`  createEffect(() => {`)
-      if (normalElems.length > 0) {
-        // Expression is always evaluated for non-conditional elements
-        lines.push(`    const __val = ${expr}`)
-        for (const elem of normalElems) {
-          const v = varSlotId(elem.slotId)
-          lines.push(`    if (_${v} && !__val?.__isSlot) _${v}.nodeValue = String(__val ?? '')`)
-        }
-        for (const elem of conditionalElems) {
-          const v = varSlotId(elem.slotId)
-          lines.push(`    const [__el_${v}] = $t(__scope, '${elem.slotId}')`)
-          lines.push(`    if (__el_${v} && !__val?.__isSlot) __el_${v}.nodeValue = String(__val ?? '')`)
-        }
-      } else {
-        // Only conditional elements — evaluate expression unconditionally
-        // to maintain reactive subscriptions even when the DOM element hasn't
-        // been created yet by insert(). Without this, the effect loses all
-        // subscriptions when the text node doesn't exist and never re-runs.
-        // Use try-catch because the expression may access a property of the
-        // conditional guard variable (e.g., prev.title where prev may be
-        // undefined when the condition is false).
-        lines.push(`    let __val`)
-        lines.push(`    try { __val = ${expr} } catch { return }`)
-        for (const elem of conditionalElems) {
-          const v = varSlotId(elem.slotId)
-          lines.push(`    const [__el_${v}] = $t(__scope, '${elem.slotId}')`)
-          lines.push(`    if (__el_${v} && !__val?.__isSlot) __el_${v}.nodeValue = String(__val ?? '')`)
-        }
-      }
-      lines.push(`  })`)
-      lines.push('')
-    }
-  }
-}
-
-/** Emit createEffect blocks for client-only expressions using comment markers. */
-export function emitClientOnlyExpressions(lines: string[], ctx: ClientJsContext): void {
-  for (const elem of ctx.clientOnlyElements) {
-    lines.push(`  // @client: ${elem.slotId}`)
-    lines.push(`  createEffect(() => {`)
-    lines.push(`    updateClientMarker(__scope, '${elem.slotId}', ${elem.expression})`)
-    lines.push(`  })`)
-    lines.push('')
-  }
-}
-
-/** Emit createEffect blocks that sync reactive attribute values (class, value, checked, etc.). */
-export function emitReactiveAttributeUpdates(lines: string[], ctx: ClientJsContext): void {
-  if (ctx.reactiveAttrs.length > 0) {
-    const attrsBySlot = new Map<string, typeof ctx.reactiveAttrs>()
-    for (const attr of ctx.reactiveAttrs) {
-      if (!attrsBySlot.has(attr.slotId)) {
-        attrsBySlot.set(attr.slotId, [])
-      }
-      attrsBySlot.get(attr.slotId)!.push(attr)
-    }
-
-    for (const [slotId, attrs] of attrsBySlot) {
-      const v = varSlotId(slotId)
-      lines.push(`  createEffect(() => {`)
-      lines.push(`    if (_${v}) {`)
-      for (const attr of attrs) {
-        // Rewrite destructured prop references to props.xxx for live reactivity.
-        // Destructured props are const-captured once; effects must read from props object.
-        const expression = rewriteDestructuredPropsInExpr(attr.expression, ctx)
-        for (const stmt of emitAttrUpdate(`_${v}`, attr.attrName, expression, attr)) {
-          lines.push(`      ${stmt}`)
-        }
-      }
-      lines.push(`    }`)
-      lines.push(`  })`)
-      lines.push('')
-    }
-  }
-}
-
-/**
- * Rewrite destructured prop names in an expression to `(props.xxx ?? default)`.
- * Only applies when the component uses destructured props (not props.xxx style).
- */
-function rewriteDestructuredPropsInExpr(expr: string, ctx: ClientJsContext): string {
-  // Skip if the component already uses props object access (not destructuring)
-  if (ctx.propsObjectName) return expr
-
-  let result = expr
-  for (const prop of ctx.propsParams) {
-    if (prop.name === 'children') continue
-    const pattern = new RegExp(`(?<![-.])\\b${prop.name}\\b`, 'g')
-    if (!pattern.test(result)) continue
-
-    const defaultVal = prop.defaultValue
-    const replacement = defaultVal
-      ? `(${PROPS_PARAM}.${prop.name} ?? ${prop.defaultContainsArrow ? `(${defaultVal})` : defaultVal})`
-      : `${PROPS_PARAM}.${prop.name}`
-    result = result.replace(new RegExp(`(?<![-.])\\b${prop.name}\\b`, 'g'), replacement)
-  }
-
-  return result
-}
-
-/**
- * Emit find() + event binding + ref callbacks + child component inits for a conditional branch.
- * Used by both emitConditionalUpdates and emitClientOnlyConditionals.
- */
-function emitBranchBindings(
-  lines: string[],
-  events: ConditionalBranchEvent[],
-  refs: ConditionalBranchRef[],
-  childComponents: ConditionalBranchChildComponent[],
-  eventNameFn: (eventName: string) => string,
-  textEffects: ConditionalBranchTextEffect[] = []
-): void {
-  const allSlotIds = new Set<string>()
-  for (const event of events) allSlotIds.add(event.slotId)
-  for (const ref of refs) allSlotIds.add(ref.slotId)
-
-  const eventsBySlot = new Map<string, ConditionalBranchEvent[]>()
-  for (const event of events) {
-    if (!eventsBySlot.has(event.slotId)) {
-      eventsBySlot.set(event.slotId, [])
-    }
-    eventsBySlot.get(event.slotId)!.push(event)
-  }
-
-  if (allSlotIds.size > 0) {
-    const slotArr = [...allSlotIds]
-    const vars = slotArr.map(id => `_${varSlotId(id)}`).join(', ')
-    const args = slotArr.map(id => `'${id}'`).join(', ')
-    lines.push(`      const [${vars}] = $(__branchScope, ${args})`)
-  }
-
-  for (const [slotId, slotEvents] of eventsBySlot) {
-    const v = varSlotId(slotId)
-    for (const event of slotEvents) {
-      const wrappedHandler = wrapHandlerInBlock(event.handler)
-      lines.push(`      if (_${v}) _${v}.addEventListener('${eventNameFn(event.eventName)}', ${wrappedHandler})`)
-    }
-  }
-
-  for (const ref of refs) {
-    const v = varSlotId(ref.slotId)
-    lines.push(`      if (_${v}) (${ref.callback})(_${v})`)
-  }
-
-  // Initialize child components created by the branch swap
-  for (let i = 0; i < childComponents.length; i++) {
-    const comp = childComponents[i]
-    const varName = `__c${i}`
-    const selectorArg = comp.slotId || comp.name
-    lines.push(`      const [${varName}] = $c(__branchScope, '${selectorArg}')`)
-    lines.push(`      if (${varName}) initChild('${comp.name}', ${varName}, ${comp.propsExpr})`)
-  }
-
-  // Emit disposable text effects scoped to this branch.
-  // These only run while the branch is active and are disposed on branch switch.
-  // Text node is resolved once (stable while branch is active) and closed over.
-  if (textEffects.length > 0) {
-    lines.push(`      const __disposers = []`)
-    for (const te of textEffects) {
-      const v = varSlotId(te.slotId)
-      lines.push(`      const [__el_${v}] = $t(__branchScope, '${te.slotId}')`)
-      lines.push(`      __disposers.push(createDisposableEffect(() => {`)
-      lines.push(`        const __val = ${te.expression}`)
-      lines.push(`        if (__el_${v} && !__val?.__isSlot) __el_${v}.nodeValue = String(__val ?? '')`)
-      lines.push(`      }))`)
-    }
-    lines.push(`      return () => __disposers.forEach(d => d())`)
-  }
-}
-
-/** Emit insert() calls for server-rendered reactive conditionals with branch configs. */
-export function emitConditionalUpdates(lines: string[], ctx: ClientJsContext): void {
-  for (const elem of ctx.conditionalElements) {
-    const whenTrueWithCond = addCondAttrToTemplate(elem.whenTrueHtml, elem.slotId)
-    const whenFalseWithCond = addCondAttrToTemplate(elem.whenFalseHtml, elem.slotId)
-
-    lines.push(`  insert(__scope, '${elem.slotId}', () => ${elem.condition}, {`)
-    lines.push(`    template: () => \`${whenTrueWithCond}\`,`)
-    lines.push(`    bindEvents: (__branchScope) => {`)
-    emitBranchBindings(lines, elem.whenTrueEvents, elem.whenTrueRefs, elem.whenTrueChildComponents, toDomEventName, elem.whenTrueTextEffects)
-    lines.push(`    }`)
-    lines.push(`  }, {`)
-    lines.push(`    template: () => \`${whenFalseWithCond}\`,`)
-    lines.push(`    bindEvents: (__branchScope) => {`)
-    emitBranchBindings(lines, elem.whenFalseEvents, elem.whenFalseRefs, elem.whenFalseChildComponents, toDomEventName, elem.whenFalseTextEffects)
-    lines.push(`    }`)
-    lines.push(`  })`)
-    lines.push('')
-  }
-}
-
-/** Emit insert() calls for client-only conditionals (not server-rendered). */
-export function emitClientOnlyConditionals(lines: string[], ctx: ClientJsContext): void {
-  for (const elem of ctx.clientOnlyConditionals) {
-    const whenTrueWithCond = addCondAttrToTemplate(elem.whenTrueHtml, elem.slotId)
-    const whenFalseWithCond = addCondAttrToTemplate(elem.whenFalseHtml, elem.slotId)
-    const rawEventName = (eventName: string) => eventName
-
-    lines.push(`  // @client conditional: ${elem.slotId}`)
-    lines.push(`  insert(__scope, '${elem.slotId}', () => ${elem.condition}, {`)
-    lines.push(`    template: () => \`${whenTrueWithCond}\`,`)
-    lines.push(`    bindEvents: (__branchScope) => {`)
-    emitBranchBindings(lines, elem.whenTrueEvents, elem.whenTrueRefs, elem.whenTrueChildComponents, rawEventName, elem.whenTrueTextEffects)
-    lines.push(`    }`)
-    lines.push(`  }, {`)
-    lines.push(`    template: () => \`${whenFalseWithCond}\`,`)
-    lines.push(`    bindEvents: (__branchScope) => {`)
-    emitBranchBindings(lines, elem.whenFalseEvents, elem.whenFalseRefs, elem.whenFalseChildComponents, rawEventName, elem.whenFalseTextEffects)
-    lines.push(`    }`)
-    lines.push(`  })`)
-    lines.push('')
-  }
-}
-
-/** Emit reconcileElements/reconcileTemplates calls for dynamic loops, and static array non-initChild updates. */
-export function emitLoopUpdates(lines: string[], ctx: ClientJsContext): void {
-  for (const elem of ctx.loopElements) {
-    if (elem.isStaticArray) {
-      // Static array initChild calls are deferred to emitStaticArrayChildInits()
-      // so that parent context providers (provideContext) run first.
-
-      // Reactive attribute effects for plain elements in static arrays.
-      // Static arrays are server-rendered once, so signal-dependent attributes
-      // (e.g., className that reads activeTag()) need createEffect to update the DOM.
-      if (!elem.childComponent && elem.childReactiveAttrs.length > 0) {
-        const v = varSlotId(elem.slotId)
-        lines.push(`  // Reactive attributes in static array children`)
-        lines.push(`  if (_${v}) {`)
-        const indexParam = elem.index || '__idx'
-        lines.push(`    ${elem.array}.forEach((${elem.param}, ${indexParam}) => {`)
-        lines.push(`      const __iterEl = _${v}.children[${indexParam}]`)
-        lines.push(`      if (__iterEl) {`)
-        // Group attrs by childSlotId to avoid duplicate const declarations
-        const attrsBySlot = new Map<string, typeof elem.childReactiveAttrs>()
-        for (const attr of elem.childReactiveAttrs) {
-          if (!attrsBySlot.has(attr.childSlotId)) {
-            attrsBySlot.set(attr.childSlotId, [])
-          }
-          attrsBySlot.get(attr.childSlotId)!.push(attr)
-        }
-        for (const [slotId, attrs] of attrsBySlot) {
-          const varName = `__t_${varSlotId(slotId)}`
-          lines.push(`        const ${varName} = qsa(__iterEl, '[bf="${slotId}"]')`)
-          lines.push(`        if (${varName}) {`)
-          for (const attr of attrs) {
-            lines.push(`          createEffect(() => {`)
-            for (const stmt of emitAttrUpdate(varName, attr.attrName, attr.expression, attr)) {
-              lines.push(`            ${stmt}`)
-            }
-            lines.push(`          })`)
-          }
-          lines.push(`        }`)
-        }
-        lines.push(`      }`)
-        lines.push(`    })`)
-        lines.push(`  }`)
-        lines.push('')
-      }
-
-      // Event delegation for plain elements in static arrays (#537)
-      // Static arrays have no data-key/bf-i markers, so walk up from target to
-      // the container's direct child and use indexOf for index lookup.
-      if (!elem.childComponent && elem.childEvents.length > 0) {
-        const v = varSlotId(elem.slotId)
-        emitLoopEventDelegation(lines, `_${v}`, elem.childEvents, (ls, ev, handlerCall, cVar) => {
-          ls.push(`      let __el = ${varSlotId(ev.childSlotId)}El`)
-          ls.push(`      while (__el.parentElement && __el.parentElement !== ${cVar}) __el = __el.parentElement`)
-          ls.push(`      if (__el.parentElement === ${cVar}) {`)
-          ls.push(`        const __idx = Array.from(${cVar}.children).indexOf(__el)`)
-          ls.push(`        const ${elem.param} = ${elem.array}[__idx]`)
-          ls.push(`        if (${elem.param}) ${handlerCall}`)
-          ls.push(`      }`)
-        })
-      }
-
-      continue
-    }
-
-    const keyFn = elem.key
-      ? `(${elem.param}${elem.index ? `, ${elem.index}` : ''}) => String(${elem.key})`
-      : 'null'
-
-    const vLoop = varSlotId(elem.slotId)
-
-    if (elem.useElementReconciliation && elem.nestedComponents?.length) {
-      emitCompositeElementReconciliation(lines, elem, keyFn, ctx)
-    } else if (elem.childComponent) {
-      const { name, props, children } = elem.childComponent
-      const propsEntries = props.map((p) => {
-        if (p.isEventHandler) {
-          return `${quotePropName(p.name)}: ${p.value}`
-        } else if (p.isLiteral) {
-          return `get ${quotePropName(p.name)}() { return ${JSON.stringify(p.value)} }`
-        } else {
-          return `get ${quotePropName(p.name)}() { return ${p.value} }`
-        }
-      })
-
-      if (children && children.length > 0) {
-        const childrenExpr = irChildrenToJsExpr(children)
-        propsEntries.push(`get children() { return ${childrenExpr} }`)
-      }
-
-      const propsExpr = propsEntries.length > 0 ? `{ ${propsEntries.join(', ')} }` : '{}'
-      const keyExpr = elem.key || '__idx'
-      const indexParam = elem.index || '__idx'
-
-      const chainedExpr = buildChainedArrayExpr(elem)
-
-      lines.push(`  createEffect(() => {`)
-      lines.push(`    reconcileElements(_${vLoop}, ${chainedExpr}, ${keyFn}, (${elem.param}, ${indexParam}) =>`)
-      lines.push(`      createComponent('${name}', ${propsExpr}, ${keyExpr})`)
-      lines.push(`    )`)
-      lines.push(`  })`)
-    } else {
-      const chainedExprTemplate = buildChainedArrayExpr(elem)
-      const indexParamTemplate = elem.index || '__idx'
-
-      lines.push(`  createEffect(() => {`)
-      lines.push(`    const __arr = ${chainedExprTemplate}`)
-      if (elem.mapPreamble) {
-        lines.push(`    const __renderItem = (${elem.param}, ${indexParamTemplate}) => { ${elem.mapPreamble}; const __tpl = document.createElement('template'); __tpl.innerHTML = \`${elem.template}\`; return __tpl.content.firstElementChild.cloneNode(true) }`)
-      } else {
-        lines.push(`    const __renderItem = (${elem.param}, ${indexParamTemplate}) => { const __tpl = document.createElement('template'); __tpl.innerHTML = \`${elem.template}\`; return __tpl.content.firstElementChild.cloneNode(true) }`)
-      }
-      // Hydration: preserve SSR elements, tag with data-key, track signals
-      lines.push(`    if (_${vLoop} && _${vLoop}.children.length > 0 && !_${vLoop}.firstElementChild?.hasAttribute('data-key')) {`)
-      lines.push(`      Array.from(_${vLoop}.children).forEach((__hChild, ${indexParamTemplate}) => {`)
-      lines.push(`        if (${indexParamTemplate} >= __arr.length) return`)
-      lines.push(`        const ${elem.param} = __arr[${indexParamTemplate}]`)
-      if (elem.key) {
-        lines.push(`        __hChild.setAttribute('data-key', String(${elem.key}))`)
-      } else {
-        lines.push(`        __hChild.setAttribute('data-key', String(${indexParamTemplate}))`)
-      }
-      lines.push(`      })`)
-      lines.push(`      if (__arr.length > 0) __renderItem(__arr[0], 0)`)
-      lines.push(`      return`)
-      lines.push(`    }`)
-      lines.push(`    reconcileElements(_${vLoop}, __arr, ${keyFn}, __renderItem)`)
-      lines.push(`  })`)
-    }
-    lines.push('')
-
-    if (!elem.childComponent && !elem.useElementReconciliation && elem.childEvents.length > 0) {
-      if (elem.key) {
-        // Dynamic keyed: find item by data-key attribute
-        const keyWithItem = elem.key.replace(new RegExp(`\\b${elem.param}\\b`, 'g'), 'item')
-        emitLoopEventDelegation(lines, `_${vLoop}`, elem.childEvents, (ls, ev, handlerCall) => {
-          if (ev.nestedLoops.length === 0) {
-            // Direct child of outer loop — single-level lookup
-            ls.push(`      const li = ${varSlotId(ev.childSlotId)}El.closest('[data-key]')`)
-            ls.push(`      if (li) {`)
-            ls.push(`        const key = li.getAttribute('data-key')`)
-            ls.push(`        const ${elem.param} = ${elem.array}.find(item => String(${keyWithItem}) === key)`)
-            ls.push(`        if (${elem.param}) ${handlerCall}`)
-            ls.push(`      }`)
-          } else {
-            // Nested loop event — multi-level data-key-N resolution
-            const evVar = varSlotId(ev.childSlotId)
-            // Resolve inner loop keys (innermost first)
-            for (const nested of ev.nestedLoops) {
-              const dataAttr = `data-key-${nested.depth}`
-              ls.push(`      const innerLi${nested.depth} = ${evVar}El.closest('[${dataAttr}]')`)
-              ls.push(`      const innerKey${nested.depth} = innerLi${nested.depth}?.getAttribute('${dataAttr}')`)
-            }
-            // Resolve outer loop key
-            ls.push(`      const outerLi = ${evVar}El.closest('[data-key]')`)
-            ls.push(`      const outerKey = outerLi?.getAttribute('data-key')`)
-            // Resolve outer loop variable
-            ls.push(`      const ${elem.param} = ${elem.array}.find(item => String(${keyWithItem}) === outerKey)`)
-            // Resolve inner loop variables via the outer param's nested array
-            for (const nested of ev.nestedLoops) {
-              const innerKeyExpr = nested.key.replace(new RegExp(`\\b${nested.param}\\b`, 'g'), 'item')
-              ls.push(`      const ${nested.param} = ${elem.param} && ${nested.array}.find(item => String(${innerKeyExpr}) === innerKey${nested.depth})`)
-            }
-            // Guard all resolved variables
-            const allParams = [elem.param, ...ev.nestedLoops.map(n => n.param)]
-            ls.push(`      if (${allParams.join(' && ')}) ${handlerCall}`)
-          }
-        })
-      } else {
-        // Dynamic non-keyed: find item by index in parent children
-        emitLoopEventDelegation(lines, `_${vLoop}`, elem.childEvents, (ls, ev, handlerCall) => {
-          ls.push(`      const li = ${varSlotId(ev.childSlotId)}El.closest('li, [bf-i]')`)
-          ls.push(`      if (li && li.parentElement) {`)
-          ls.push(`        const idx = Array.from(li.parentElement.children).indexOf(li)`)
-          ls.push(`        const ${elem.param} = ${elem.array}[idx]`)
-          ls.push(`        if (${elem.param}) ${handlerCall}`)
-          ls.push(`      }`)
-        })
-      }
-    }
-  }
-}
-
-/**
- * Emit reconcileElements with composite rendering for dynamic loops whose
- * native-element body contains child components.
- *
- * Generates:
- * 1. A hydration pass that tags SSR elements with data-key and initializes
- *    child components via initChild() + sets up events via addEventListener
- * 2. A createEffect with reconcileElements where renderItem creates elements
- *    from a placeholder template, replaces placeholders with createComponent(),
- *    and attaches events directly.
- *
- * Events and components at different nesting levels are handled separately:
- * - Outer level (loopDepth=0): direct querySelector on the item element
- * - Inner level (loopDepth>0): iterate inner array, find elements by data-key-N
- */
-function emitCompositeElementReconciliation(
-  lines: string[],
-  elem: LoopElement,
-  keyFn: string,
-  _ctx: ClientJsContext,
-): void {
-  const vLoop = varSlotId(elem.slotId)
-  const chainedExpr = buildChainedArrayExpr(elem)
-  const indexParam = elem.index || '__idx'
-
-  // Helper: build props expression for a nested component
-  const buildPropsExpr = (comp: { props: Array<{ name: string; value: string; isEventHandler: boolean; isLiteral: boolean }>, children?: import('../types').IRNode[] }): string => {
-    const entries = comp.props.map((p) => {
-      if (p.isEventHandler) {
-        return `${quotePropName(p.name)}: ${p.value}`
-      } else if (p.isLiteral) {
-        return `${quotePropName(p.name)}: ${JSON.stringify(p.value)}`
-      } else {
-        return `get ${quotePropName(p.name)}() { return ${p.value} }`
-      }
-    })
-    if ('children' in comp && Array.isArray(comp.children) && comp.children.length > 0) {
-      const childrenExpr = irChildrenToJsExpr(comp.children)
-      entries.push(`get children() { return ${childrenExpr} }`)
-    }
-    return entries.length > 0 ? `{ ${entries.join(', ')} }` : '{}'
-  }
-
-  const nestedComps = elem.nestedComponents!
-
-  // Separate components and events by nesting level
-  const outerComps = nestedComps.filter(c => !c.loopDepth || c.loopDepth === 0)
-  const innerComps = nestedComps.filter(c => (c.loopDepth ?? 0) > 0)
-  const outerEvents = elem.childEvents.filter(ev => ev.nestedLoops.length === 0)
-  const innerEvents = elem.childEvents.filter(ev => ev.nestedLoops.length > 0)
-
-  // Extract inner loop info: prefer IR-derived innerLoops, fall back to event nesting info
-  const innerLoopInfo = elem.innerLoops?.[0]
-    ?? (innerEvents.length > 0 ? innerEvents[0].nestedLoops[0] : null)
-
-  // Helper: emit event handler setup
-  const emitEventSetup = (ls: string[], indent: string, elVar: string, ev: LoopChildEvent): void => {
-    const handler = ev.handler.trim().startsWith('(') || ev.handler.trim().startsWith('function')
-      ? `(${ev.handler})(e)`
-      : ev.handler
-    ls.push(`${indent}{ const __e = qsa(${elVar}, '[bf="${ev.childSlotId}"]'); if (__e) __e.addEventListener('${toDomEventName(ev.eventName)}', (e) => { ${handler} }) }`)
-  }
-
-  // Helper: emit renderItem body (shared between hydration tracking and reconciliation)
-  const emitRenderItemBody = (ls: string[], indent: string): void => {
-    ls.push(`${indent}const __tpl = document.createElement('template')`)
-    if (elem.mapPreamble) {
-      ls.push(`${indent}${elem.mapPreamble}`)
-    }
-    ls.push(`${indent}__tpl.innerHTML = \`${elem.template}\``)
-    ls.push(`${indent}const __el = __tpl.content.firstElementChild.cloneNode(true)`)
-
-    // Replace outer-level component placeholders
-    for (const comp of outerComps) {
-      const phId = comp.slotId || comp.name
-      const propsExpr = buildPropsExpr(comp)
-      const keyProp = comp.props.find(p => p.name === 'key')
-      const keyArg = keyProp ? `, ${keyProp.value}` : ''
-      ls.push(`${indent}{ const __ph = __el.querySelector('[data-bf-ph="${phId}"]'); if (__ph) __ph.replaceWith(createComponent('${comp.name}', ${propsExpr}${keyArg})) }`)
-    }
-
-    // Set up outer-level events
-    for (const ev of outerEvents) {
-      emitEventSetup(ls, indent, '__el', ev)
-    }
-
-    // Handle inner loop: iterate array, find elements by data-key-N
-    if (innerLoopInfo && (innerComps.length > 0 || innerEvents.length > 0)) {
-      const inner = innerLoopInfo
-      ls.push(`${indent}// Initialize inner loop components and events`)
-      ls.push(`${indent}${inner.array}.forEach((${inner.param}) => {`)
-      if (inner.key) {
-        ls.push(`${indent}  const __innerEl = __el.querySelector('[data-key-${inner.depth}="' + ${inner.key} + '"]')`)
-      } else {
-        ls.push(`${indent}  const __innerEl = null`)
-      }
-      ls.push(`${indent}  if (!__innerEl) return`)
-      for (const comp of innerComps) {
-        const phId = comp.slotId || comp.name
-        const propsExpr = buildPropsExpr(comp)
-        const keyProp = comp.props.find(p => p.name === 'key')
-        const keyArg = keyProp ? `, ${keyProp.value}` : ''
-        ls.push(`${indent}  { const __ph = __innerEl.querySelector('[data-bf-ph="${phId}"]'); if (__ph) __ph.replaceWith(createComponent('${comp.name}', ${propsExpr}${keyArg})) }`)
-      }
-      for (const ev of innerEvents) {
-        emitEventSetup(ls, `${indent}  `, '__innerEl', ev)
-      }
-      ls.push(`${indent}})`)
-    }
-
-    ls.push(`${indent}return __el`)
-  }
-
-  // Single createEffect with hydration-aware first run
-  // Pattern: first run preserves SSR content and calls renderItem once for signal tracking.
-  // Subsequent runs call reconcileElements normally.
-  lines.push(`  createEffect(() => {`)
-  lines.push(`    const __arr = ${chainedExpr}`)
-  lines.push(`    const __renderItem = (${elem.param}, ${indexParam}) => {`)
-  emitRenderItemBody(lines, '      ')
-  lines.push(`    }`)
-  lines.push('')
-  lines.push(`    // Hydration: preserve SSR elements, init components/events, track signals`)
-  lines.push(`    if (_${vLoop} && _${vLoop}.children.length > 0 && !_${vLoop}.firstElementChild?.hasAttribute('data-key')) {`)
-  lines.push(`      Array.from(_${vLoop}.children).forEach((__hChild, ${indexParam}) => {`)
-  lines.push(`        if (${indexParam} >= __arr.length) return`)
-  lines.push(`        const ${elem.param} = __arr[${indexParam}]`)
-  if (elem.key) {
-    lines.push(`        __hChild.setAttribute('data-key', String(${elem.key}))`)
-  } else {
-    lines.push(`        __hChild.setAttribute('data-key', String(${indexParam}))`)
-  }
-  // Initialize outer-level child components in SSR markup
-  // Use both suffix match (for components with slotSuffix in bf-s, e.g. ~Badge_hash_s2)
-  // and prefix match (for components without slotSuffix, e.g. ~Badge_hash) as SSR
-  // renderChild() may not include slotSuffix in the generated bf-s attribute.
-  for (const comp of outerComps) {
-    const selector = comp.slotId
-      ? `[bf-s$="_${comp.slotId}"], [bf-s^="~${comp.name}_"], [bf-s^="${comp.name}_"]`
-      : `[bf-s^="~${comp.name}_"], [bf-s^="${comp.name}_"]`
-    const propsExpr = buildPropsExpr(comp)
-    lines.push(`        { const __c = __hChild.querySelector('${selector}'); if (__c) initChild('${comp.name}', __c, ${propsExpr}) }`)
-  }
-  // Set up outer-level events on SSR elements
-  for (const ev of outerEvents) {
-    emitEventSetup(lines, '        ', '__hChild', ev)
-  }
-  // Handle inner loop items in SSR
-  // SSR HTML doesn't have data-key-N (Hono JSX strips key props), so use
-  // container children index. Tag with data-key-N for future querySelector lookups.
-  if (innerLoopInfo && (innerComps.length > 0 || innerEvents.length > 0)) {
-    const inner = innerLoopInfo
-    const containerSelector = inner.containerSlotId ? `'[bf="${inner.containerSlotId}"]'` : 'null'
-    lines.push(`        { const __ic = ${containerSelector !== 'null' ? `__hChild.querySelector(${containerSelector})` : '__hChild'}`)
-    lines.push(`        if (__ic) ${inner.array}.forEach((${inner.param}, __innerIdx) => {`)
-    lines.push(`          const __innerEl = __ic.children[__innerIdx]`)
-    lines.push(`          if (!__innerEl) return`)
-    if (inner.key) {
-      lines.push(`          __innerEl.setAttribute('data-key-${inner.depth}', String(${inner.key}))`)
-    }
-    for (const comp of innerComps) {
-      const selector = comp.slotId
-        ? `[bf-s$="_${comp.slotId}"], [bf-s^="~${comp.name}_"], [bf-s^="${comp.name}_"]`
-        : `[bf-s^="~${comp.name}_"], [bf-s^="${comp.name}_"]`
-      const propsExpr = buildPropsExpr(comp)
-      lines.push(`          { const __c = __innerEl.querySelector('${selector}'); if (__c) initChild('${comp.name}', __c, ${propsExpr}) }`)
-    }
-    for (const ev of innerEvents) {
-      emitEventSetup(lines, '          ', '__innerEl', ev)
-    }
-    lines.push(`        }) }`)
-  }
-  lines.push(`      })`)
-  lines.push(`      // Call renderItem once to track signal dependencies (template reads signals)`)
-  lines.push(`      if (__arr.length > 0) __renderItem(__arr[0], 0)`)
-  lines.push(`      return`)
-  lines.push(`    }`)
-  lines.push('')
-  // Blur active element before reconciliation to avoid syncElementState issues.
-  // Composite elements have duplicate internal slot IDs (e.g., multiple Badge components
-  // each with bf="s0") which cause syncElementState to overwrite text incorrectly.
-  // Also, syncElementState can't handle conditional structure changes (comment → element).
-  lines.push(`    if (_${vLoop}?.contains(document.activeElement)) document.activeElement?.blur()`)
-  lines.push(`    reconcileElements(_${vLoop}, __arr, ${keyFn}, __renderItem)`)
-  lines.push(`  })`)
-}
-
-/**
- * Callback that emits the item-lookup lines inside a loop event delegation handler.
- * Called once per event after target.closest() matched.
- */
-type ItemLookupEmitter = (
-  lines: string[],
-  ev: LoopChildEvent,
-  handlerCall: string,
-  containerVar: string,
-) => void
-
-/** Non-bubbling events that require addEventListener with capture for delegation. */
-const NON_BUBBLING_EVENTS = new Set(['blur', 'focus', 'load', 'unload'])
-
-/**
- * Emit event delegation for child events inside a loop (static or dynamic).
- * The shared shell (event grouping, closest matching, handler call construction)
- * is handled here; the strategy-specific item-lookup is injected via callback.
- */
-function emitLoopEventDelegation(
-  lines: string[],
-  containerVar: string,
-  childEvents: LoopChildEvent[],
-  emitItemLookup: ItemLookupEmitter,
-): void {
-  const eventsByName = new Map<string, LoopChildEvent[]>()
-  for (const ev of childEvents) {
-    if (!eventsByName.has(ev.eventName)) {
-      eventsByName.set(ev.eventName, [])
-    }
-    eventsByName.get(ev.eventName)!.push(ev)
-  }
-
-  for (const [eventName, events] of eventsByName) {
-    const useCapture = NON_BUBBLING_EVENTS.has(eventName)
-    if (useCapture) {
-      lines.push(`  if (${containerVar}) ${containerVar}.addEventListener('${eventName}', (e) => {`)
-    } else {
-      lines.push(`  if (${containerVar}) ${containerVar}.addEventListener('${toDomEventName(eventName)}', (e) => {`)
-    }
-    lines.push(`    const target = e.target`)
-    for (const ev of events) {
-      const childVar = varSlotId(ev.childSlotId)
-      lines.push(`    const ${childVar}El = target.closest('[bf="${ev.childSlotId}"]')`)
-      lines.push(`    if (${childVar}El) {`)
-      const handlerCall = ev.handler.trim().startsWith('(') || ev.handler.trim().startsWith('function')
-        ? `(${ev.handler})(e)`
-        : ev.handler
-      emitItemLookup(lines, ev, handlerCall, containerVar)
-      lines.push(`      return`)
-      lines.push(`    }`)
-    }
-    if (useCapture) {
-      lines.push(`  }, true)`)
-    } else {
-      lines.push(`  })`)
-    }
-    lines.push('')
-  }
-}
-
 /** Emit applyRestAttrs() calls for HTML elements with unresolved spread attrs. */
 export function emitRestAttrApplications(lines: string[], ctx: ClientJsContext): void {
   for (const elem of ctx.restAttrElements) {
@@ -921,97 +227,6 @@ export function emitEventHandlers(
         lines.push(`  if (_${v}) _${v}.addEventListener('${eventName}', ${wrappedHandler})`)
       }
     }
-  }
-}
-
-/** Emit createEffect to update component element attributes when signal/memo values change. */
-export function emitReactivePropBindings(lines: string[], ctx: ClientJsContext): void {
-  if (ctx.reactiveProps.length > 0) {
-    lines.push('')
-    lines.push(`  // Reactive prop bindings`)
-    lines.push(`  createEffect(() => {`)
-
-    const propsBySlot = new Map<string, typeof ctx.reactiveProps>()
-    for (const prop of ctx.reactiveProps) {
-      if (!propsBySlot.has(prop.slotId)) {
-        propsBySlot.set(prop.slotId, [])
-      }
-      propsBySlot.get(prop.slotId)!.push(prop)
-    }
-
-    for (const [slotId, props] of propsBySlot) {
-      const v = varSlotId(slotId)
-      lines.push(`    if (_${v}) {`)
-      for (const prop of props) {
-        const value = `${prop.expression}()`
-        if (prop.propName === 'selected') {
-          if (prop.componentName === 'TabsContent') {
-            lines.push(`      _${v}.setAttribute('data-state', ${value} ? 'active' : 'inactive')`)
-            lines.push(`      if (${value}) {`)
-            lines.push(`        _${v}.classList.remove('hidden')`)
-            lines.push(`      } else {`)
-            lines.push(`        _${v}.classList.add('hidden')`)
-            lines.push(`      }`)
-          } else {
-            // Update data-state and aria-selected attributes.
-            // Visual styling is driven by CSS data-[state=active/inactive]: selectors.
-            lines.push(`      _${v}.setAttribute('aria-selected', String(${value}))`)
-            lines.push(`      _${v}.setAttribute('data-state', ${value} ? 'active' : 'inactive')`)
-            lines.push(`      _${v}.setAttribute('tabindex', ${value} ? '0' : '-1')`)
-          }
-        // Use DOM property assignment for value and boolean attrs.
-        // setAttribute('value', x) only sets the initial HTML attribute; after user
-        // interaction the DOM property diverges, so .value = x is required.
-        // Boolean attrs (disabled, checked, etc.) treat any attribute presence as
-        // truthy, so setAttribute('disabled', 'false') still disables the element.
-        } else if (prop.propName === 'value') {
-          lines.push(`      const __val = String(${value})`)
-          lines.push(`      if (_${v}.value !== __val) _${v}.value = __val`)
-        } else if (isBooleanAttr(prop.propName)) {
-          lines.push(`      _${v}.${prop.propName} = !!(${value})`)
-        } else {
-          lines.push(`      _${v}.setAttribute('${prop.propName}', String(${value}))`)
-        }
-      }
-      lines.push(`    }`)
-    }
-
-    lines.push(`  })`)
-  }
-}
-
-/** Emit createEffect to update child component DOM attributes when parent props change. */
-export function emitReactiveChildProps(lines: string[], ctx: ClientJsContext): void {
-  if (ctx.reactiveChildProps.length > 0) {
-    lines.push('')
-    lines.push(`  // Reactive child component props`)
-    lines.push(`  createEffect(() => {`)
-
-    const propsByComponent = new Map<string, typeof ctx.reactiveChildProps>()
-    for (const prop of ctx.reactiveChildProps) {
-      const key = `${prop.componentName}_${prop.slotId ?? '__scope'}`
-      if (!propsByComponent.has(key)) {
-        propsByComponent.set(key, [])
-      }
-      propsByComponent.get(key)!.push(prop)
-    }
-
-    for (const [, props] of propsByComponent) {
-      const first = props[0]
-      const varSuffix = first.slotId ? varSlotId(first.slotId).replace(/-/g, '_') : first.componentName
-      const varName = `__${first.componentName}_${varSuffix}El`
-      const selectorArg = first.slotId ? first.slotId : first.componentName
-      lines.push(`    const [${varName}] = $c(__scope, '${selectorArg}')`)
-      lines.push(`    if (${varName}) {`)
-      for (const prop of props) {
-        for (const stmt of emitAttrUpdate(varName, prop.attrName, prop.expression, prop)) {
-          lines.push(`      ${stmt}`)
-        }
-      }
-      lines.push(`    }`)
-    }
-
-    lines.push(`  })`)
   }
 }
 
@@ -1136,385 +351,4 @@ export function emitStaticArrayChildInits(lines: string[], ctx: ClientJsContext)
       }
     }
   }
-}
-
-// JavaScript built-in identifiers that are always available at any scope
-const JS_BUILTINS = new Set([
-  'true', 'false', 'null', 'undefined', 'NaN', 'Infinity',
-  'typeof', 'instanceof', 'void', 'delete', 'new', 'in', 'of',
-  'this', 'super', 'return', 'throw', 'if', 'else',
-  'for', 'while', 'do', 'switch', 'case', 'break', 'continue',
-  'try', 'catch', 'finally', 'yield', 'await', 'async',
-  'let', 'const', 'var', 'function', 'class',
-  'Math', 'JSON', 'Object', 'Array', 'String', 'Number', 'Boolean',
-  'Date', 'RegExp', 'Map', 'Set', 'WeakMap', 'WeakSet', 'Promise',
-  'Error', 'TypeError', 'RangeError', 'SyntaxError',
-  'console', 'window', 'document', 'globalThis', 'navigator',
-  'parseInt', 'parseFloat', 'isNaN', 'isFinite',
-  'encodeURIComponent', 'decodeURIComponent', 'encodeURI', 'decodeURI',
-  'setTimeout', 'clearTimeout', 'setInterval', 'clearInterval',
-  'requestAnimationFrame', 'cancelAnimationFrame',
-  'Symbol', 'Proxy', 'Reflect', 'BigInt',
-])
-
-/**
- * Check that an expression value only references identifiers known within
- * the component scope (or JavaScript built-ins). Returns false if the value
- * contains references to file-scope variables that won't be available in
- * the generated client JS module scope.
- *
- * Uses pre-computed freeIdentifiers from the analyzer phase (ConstantInfo.freeIdentifiers).
- */
-function valueOnlyUsesKnownNames(freeIds: Set<string>, knownNames: Set<string>): boolean {
-  for (const id of freeIds) {
-    if (JS_BUILTINS.has(id)) continue
-    if (knownNames.has(id)) continue
-    return false
-  }
-  return true
-}
-
-/**
- * Resolve chained references within a constants map.
- * If constant A references constant B, replace B's name in A's value with B's resolved value.
- * Uses pre-computed freeIdentifiers to skip unnecessary regex replacements.
- */
-export function resolveChainedRefs(constants: Map<string, string>, freeIdsMap?: Map<string, Set<string>>): void {
-  let changed = true
-  const maxIterations = constants.size + 1
-  let iteration = 0
-  while (changed && iteration < maxIterations) {
-    changed = false
-    iteration++
-    for (const [constName, constValue] of constants) {
-      // String literal values (single/double quoted) cannot contain variable references.
-      // Skip them to avoid corrupting CSS class names like "size-4" when a constant
-      // named "size" exists (the regex would falsely match "size" in "size-4").
-      const trimmed = constValue.trim()
-      if ((trimmed.startsWith("'") && trimmed.endsWith("'")) ||
-          (trimmed.startsWith('"') && trimmed.endsWith('"'))) {
-        continue
-      }
-
-      // Protect string literals within compound values (e.g., Record objects
-      // containing 'size-4') from regex-based identifier replacement
-      const { protect, restore } = createStringProtector()
-      let newValue = protect(constValue)
-      const freeIds = freeIdsMap?.get(constName)
-      for (const [otherName, otherValue] of constants) {
-        if (otherName === constName) continue
-        // Use freeIdentifiers to skip constants that are definitely not referenced.
-        // On subsequent iterations the value may have been expanded beyond what
-        // freeIdentifiers tracks, so fall through to regex when freeIds is unavailable
-        // or when the iteration is > 1 (values have been mutated).
-        if (freeIds && iteration === 1 && !freeIds.has(otherName)) continue
-        const replaced = newValue.replace(new RegExp(`(?<![-.])\\b${otherName}\\b`, 'g'), `(${protect(otherValue)})`)
-        if (replaced !== newValue) {
-          newValue = replaced
-          changed = true
-        }
-      }
-      newValue = restore(newValue)
-      if (newValue !== constValue) {
-        constants.set(constName, newValue)
-      }
-    }
-  }
-}
-
-/**
- * Build the inlinable constants map and unsafe local names set from a context.
- * Extracted for reuse by both emitRegistrationAndHydration and generateTemplateOnlyMount (#435).
- */
-export function buildInlinableConstants(ctx: ClientJsContext): {
-  inlinableConstants: Map<string, string>
-  unsafeLocalNames: Set<string>
-} {
-  const inlinableConstants = new Map<string, string>()
-  const unsafeLocalNames = new Set<string>()
-
-  const signalGetters = new Set(ctx.signals.map(s => s.getter))
-  const signalSetters = new Set(ctx.signals.map(s => s.setter))
-  const memoNames = new Set(ctx.memos.map(m => m.name))
-
-  const componentScopeNames = new Set<string>()
-  for (const c of ctx.localConstants) componentScopeNames.add(c.name)
-  for (const s of ctx.signals) { componentScopeNames.add(s.getter); componentScopeNames.add(s.setter) }
-  for (const m of ctx.memos) componentScopeNames.add(m.name)
-  for (const f of ctx.localFunctions) componentScopeNames.add(f.name)
-  for (const p of ctx.propsParams) componentScopeNames.add(p.name)
-  if (ctx.propsObjectName) componentScopeNames.add(ctx.propsObjectName)
-
-  for (const fn of ctx.localFunctions) {
-    // Module-level functions that don't reference component internals
-    // are emitted at module scope, so they are available in the template.
-    if (fn.isModule && !bodyReferencesComponentScope(fn.body, componentScopeNames)) continue
-    unsafeLocalNames.add(fn.name)
-  }
-
-  for (const constant of ctx.localConstants) {
-    if (constant.isJsx) continue  // Inlined at IR level (#547)
-    if (!constant.value) {
-      // `let x` with no initializer — not safe for template inlining
-      unsafeLocalNames.add(constant.name)
-      continue
-    }
-    const trimmedValue = constant.value.trim()
-
-    // Use AST-derived flag instead of string inspection
-    if (constant.containsArrow) {
-      unsafeLocalNames.add(constant.name)
-      continue
-    }
-
-    // Use AST-derived flag instead of regex
-    if (constant.systemConstructKind) {
-      continue
-    }
-
-    // Use pre-computed freeIdentifiers instead of regex
-    let dependsOnReactive = false
-    const freeIds = constant.freeIdentifiers
-    if (freeIds) {
-      for (const sigName of signalGetters) {
-        if (freeIds.has(sigName)) { dependsOnReactive = true; break }
-      }
-      if (!dependsOnReactive) {
-        for (const setterName of signalSetters) {
-          if (freeIds.has(setterName)) { dependsOnReactive = true; break }
-        }
-      }
-      if (!dependsOnReactive) {
-        for (const mName of memoNames) {
-          if (freeIds.has(mName)) { dependsOnReactive = true; break }
-        }
-      }
-    }
-
-    if (dependsOnReactive) {
-      unsafeLocalNames.add(constant.name)
-      continue
-    }
-
-    if (!valueOnlyUsesKnownNames(constant.freeIdentifiers!, componentScopeNames)) {
-      unsafeLocalNames.add(constant.name)
-      continue
-    }
-
-    inlinableConstants.set(constant.name, trimmedValue)
-  }
-
-  // Build freeIdentifiers lookup for resolveChainedRefs
-  const freeIdsMap = new Map<string, Set<string>>()
-  for (const constant of ctx.localConstants) {
-    if (constant.freeIdentifiers) {
-      freeIdsMap.set(constant.name, constant.freeIdentifiers)
-    }
-  }
-
-  resolveChainedRefs(inlinableConstants, freeIdsMap)
-
-  // Demote constants whose value still references an unsafe name.
-  // Use freeIdentifiers from the original constant for initial check,
-  // then fall back to checking the resolved value (which may have been
-  // expanded by resolveChainedRefs and no longer matches the original freeIdentifiers).
-  const toRemove: string[] = []
-  for (const [constName, constValue] of inlinableConstants) {
-    const constFreeIds = freeIdsMap.get(constName)
-    let isUnsafe = false
-    if (constFreeIds) {
-      for (const unsafeName of unsafeLocalNames) {
-        if (constFreeIds.has(unsafeName)) { isUnsafe = true; break }
-      }
-    }
-    // After chained resolution, the constValue may contain identifiers
-    // from inlined constants that are themselves unsafe. Fall back to
-    // regex for the resolved value since freeIdentifiers reflect the
-    // original source, not the resolved string.
-    if (!isUnsafe) {
-      for (const unsafeName of unsafeLocalNames) {
-        if (new RegExp(`\\b${unsafeName}\\b`).test(constValue)) {
-          isUnsafe = true
-          break
-        }
-      }
-    }
-    if (isUnsafe) {
-      toRemove.push(constName)
-    }
-  }
-  for (const removeName of toRemove) {
-    inlinableConstants.delete(removeName)
-    unsafeLocalNames.add(removeName)
-  }
-
-  return { inlinableConstants, unsafeLocalNames }
-}
-
-/**
- * Build signal and memo maps for CSR template generation.
- * Signal map: getter name → initial value expression
- * Memo map: memo name → computation expression with signal calls replaced by initial values
- */
-export function buildSignalAndMemoMaps(ctx: ClientJsContext): {
-  signalMap: Map<string, string>
-  memoMap: Map<string, string>
-} {
-  const propsName = ctx.propsObjectName ?? 'props'
-  const propsPrefix = `${propsName}.`
-
-  const signalMap = new Map<string, string>()
-  for (const signal of ctx.signals) {
-    let initialValue = signal.initialValue
-    // Normalize custom props object name to PROPS_PARAM and add default fallback
-    // to match emitSignalsAndMemos() behavior (prevents undefined rendering in CSR)
-    if (ctx.propsObjectName && initialValue.startsWith(propsPrefix)) {
-      const propRef = `${PROPS_PARAM}.` + initialValue.slice(propsPrefix.length)
-      if (!initialValue.includes('??')) {
-        initialValue = `${propRef} ?? ${inferDefaultValue(signal.type)}`
-      } else {
-        initialValue = propRef
-      }
-    } else if (initialValue.startsWith('props.') && !initialValue.includes('??')) {
-      initialValue = `${PROPS_PARAM}.${initialValue.slice('props.'.length)} ?? ${inferDefaultValue(signal.type)}`
-    }
-    signalMap.set(signal.getter, initialValue)
-  }
-
-  const memoMap = new Map<string, string>()
-  for (const memo of ctx.memos) {
-    let expr = memo.computation
-    // Extract the function body from arrow function: () => count() * 2 → count() * 2
-    // Supports both expression arrows `() => expr` and block arrows `() => { return expr }`
-    const arrowMatch = expr.match(/^\(\)\s*=>\s*(.+)$/s)
-    if (arrowMatch) {
-      const body = arrowMatch[1].trim()
-      if (body.startsWith('{')) {
-        // Block body: extract return expression
-        const returnMatch = body.match(/return\s+(.+?)\s*[;}]?\s*}$/)
-        expr = returnMatch ? returnMatch[1] : expr
-      } else {
-        expr = body
-      }
-    }
-    // Replace signal getter calls with initial values
-    for (const [getter, initial] of signalMap) {
-      expr = expr.replace(new RegExp(`\\b${getter}\\(\\)`, 'g'), `(${initial})`)
-    }
-    memoMap.set(memo.name, expr)
-  }
-
-  // Resolve chained memo references: if memo A references memo B(),
-  // replace B() with B's resolved computation expression.
-  let changed = true
-  const maxIter = memoMap.size + 1
-  let iter = 0
-  while (changed && iter < maxIter) {
-    changed = false
-    iter++
-    for (const [memoName, memoExpr] of memoMap) {
-      let newExpr = memoExpr
-      for (const [otherName, otherExpr] of memoMap) {
-        if (otherName === memoName) continue
-        const replaced = newExpr.replace(new RegExp(`\\b${otherName}\\(\\)`, 'g'), `(${otherExpr})`)
-        if (replaced !== newExpr) {
-          newExpr = replaced
-          changed = true
-        }
-      }
-      if (newExpr !== memoExpr) {
-        memoMap.set(memoName, newExpr)
-      }
-    }
-  }
-
-  return { signalMap, memoMap }
-}
-
-/**
- * Build an expanded inlinable constants map for CSR template generation.
- * Re-promotes constants that were demoted to unsafeLocalNames by resolving
- * signal/memo call expressions with their initial values.
- */
-export function buildCsrInlinableConstants(
-  ctx: ClientJsContext,
-  inlinableConstants: Map<string, string>,
-  unsafeLocalNames: Set<string>,
-  signalMap: Map<string, string>,
-  memoMap: Map<string, string>,
-): Map<string, string> {
-  const csrInlinableConstants = new Map(inlinableConstants)
-  for (const constant of ctx.localConstants) {
-    if (unsafeLocalNames.has(constant.name) && constant.value && !constant.containsArrow) {
-      let value = constant.value.trim()
-      for (const [getter, initial] of signalMap) {
-        value = value.replace(new RegExp(`\\b${getter}\\(\\)`, 'g'), `(${initial})`)
-      }
-      for (const [memoName, computation] of memoMap) {
-        value = value.replace(new RegExp(`\\b${memoName}\\(\\)`, 'g'), `(${computation})`)
-      }
-      if (!/\b\w+\(\)/.test(value)) {
-        csrInlinableConstants.set(constant.name, value)
-      }
-    }
-  }
-  resolveChainedRefs(csrInlinableConstants)
-  return csrInlinableConstants
-}
-
-/** Emit hydrate() call that registers component, template, and hydrates. */
-/**
- * Generate the closing brace for the init function and the hydrate() call.
- * Returns the hydrate line separately so it can be appended after props renaming,
- * preventing double-replacement of props references in template expressions.
- */
-export function emitRegistrationAndHydration(
-  lines: string[],
-  ctx: ClientJsContext,
-  _ir: ComponentIR,
-): string {
-  const name = ctx.componentName
-
-  lines.push(`}`)
-  lines.push('')
-
-  const propNamesForTemplate = new Set(ctx.propsParams.map((p) => p.name))
-  const { inlinableConstants, unsafeLocalNames } = buildInlinableConstants(ctx)
-
-  // Build rest spread names: these are rest/props spreads handled by applyRestAttrs, not spreadAttrs
-  const restSpreadNames = new Set<string>()
-  if (ctx.restPropsName) restSpreadNames.add(ctx.restPropsName)
-  if (ctx.propsObjectName) restSpreadNames.add(ctx.propsObjectName)
-
-  const isCommentScope = (_ir.root.type === 'fragment'
-    && (_ir.root as IRFragment).needsScopeComment)
-    || _ir.root.type === 'component'
-
-  // Build ComponentDef object for hydrate()
-  const defParts: string[] = [`init: init${name}`]
-  if (canGenerateStaticTemplate(_ir.root, propNamesForTemplate, inlinableConstants, unsafeLocalNames)) {
-    const templateHtml = irToComponentTemplate(_ir.root, propNamesForTemplate, inlinableConstants, restSpreadNames, ctx.propsObjectName)
-    if (templateHtml) {
-      defParts.push(`template: (${PROPS_PARAM}) => \`${templateHtml}\``)
-    }
-  } else {
-    // CSR fallback: emit for all components that can't generate static templates.
-    // Components may be imported and used in conditional branches in other files,
-    // where renderChild() needs a registered template to render HTML correctly.
-    const { signalMap, memoMap } = buildSignalAndMemoMaps(ctx)
-    const csrInlinableConstants = buildCsrInlinableConstants(ctx, inlinableConstants, unsafeLocalNames, signalMap, memoMap)
-
-    const templateHtml = generateCsrTemplate(
-      _ir.root, propNamesForTemplate, csrInlinableConstants, signalMap, memoMap, undefined, restSpreadNames, ctx.propsObjectName
-    )
-    if (templateHtml) {
-      defParts.push(`template: (${PROPS_PARAM}) => \`${templateHtml}\``)
-    }
-  }
-  // No else: top-level-only components skip template entirely (save bytes)
-  if (isCommentScope) {
-    defParts.push('comment: true')
-  }
-
-  return `hydrate('${name}', { ${defParts.join(', ')} })`
 }
